requirements.txt
Flask
mysql-connector-python
hashids
inflect
PyJWT